requirements.txt
arxiv>=2.1.0
pymupdf>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
matplotlib>=3.7.0
numpy>=1.24.0
pandas>=1.5.0
scikit-learn>=1.2.0
Pillow>=9.4.0
tabulate>=0.9.0
psutil>=5.9.0
requests>=2.28.0
tqdm>=4.65.0
anthropic>=0.39.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
sse-starlette>=1.8.0
